GCP_cloud_run/intersection_geojson/requirements.txt
flask==2.0.2
flask_restful==0.3.9
marshmallow==3.14.1
gunicorn==20.1.0
DateTime==4.3
google-cloud-bigquery==2.34.2
python-dateutil==2.8.2
pytz==2021.3